Extract `FileScanConfig` into its own module
